# server/src/main/resources/static/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "An API for the Spring2885 FrontEnd.",
    "version" : "0.0.1",
    "title" : "NDNU Community Access API",
    "contact" : {
      "name" : "Spring2885 Backend Development Team",
      "url" : "https://github.com/spring2885/backend",
      "email" : "team@spring-2885.org"
    }
  },
...
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/api/v1/approvals/list" : {
      "get" : {
        "tags" : [ "Approvals" ],
        "summary" : "Lists approvals.",
        "description" : "Lists all approval requests. Optionally filtered by state and type.",
        "parameters" : [ {
          "name" : "state",
          "in" : "query",
          "description" : "The state of the request. Either *open* or *closed*",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "type",
          "in" : "query",
          "description" : "The type of the request. Either *FACULTY* or *ABUSE*",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Submits verdict."
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
    "/api/v1/approvals/request/abuse" : {
      "post" : {
        "tags" : [ "Approvals" ],
        "summary" : "Flags an item as abuse.",
        "description" : "Submits a request to flag an item as abuse.",
        "parameters" : [ {
          "in" : "body",
          "name" : "AbuseRequest",
          "description" : "The faculty request.",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/AbuseRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Submits abuse request."
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
    "/api/v1/approvals/request/faculty" : {
      "post" : {
        "tags" : [ "Approvals" ],
        "summary" : "Requests approval as faculty",
        "description" : "Submits a request for an admin to approve this member as a faculty.",
        "parameters" : [ {
          "in" : "body",
          "name" : "FacultyRequest",
          "description" : "The faculty request.",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/FacultyRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Submits faculty request."
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
    "/api/v1/approvals/verdict" : {
      "post" : {
        "tags" : [ "Approvals" ],
        "summary" : "Submits a verdict on an approval request.",
        "description" : "Submits a verdict on a previously submitted approval request.",
        "parameters" : [ {
          "in" : "body",
          "name" : "Verdict",
          "description" : "The verdict.",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Verdict"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Submits verdict."
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/persontype" : {
      "get" : {
        "tags" : [ "PersonType" ],
        "summary" : "Fetches all persontype entries.\n",
        "description" : "Fetches all 'person type' objects. up to a 'size' if specified,\notherwise fetches all objects.\n",
        "parameters" : [ {
          "name" : "size",
          "in" : "query",
          "description" : "Maximum number of items to fetch.\nIf unset then all items are fetched.\n",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Gets persontypes",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PersonType"
              }
            }
          },
          "500" : {
            "description" : "Error getting persontypes"
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "PersonType" ],
        "summary" : "Creates a new persontype.",
        "description" : "Creates a new persontype entry.",
        "parameters" : [ {
          "in" : "body",
          "name" : "PersonType",
          "description" : "The persontype you want to post",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/PersonType"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Creates a new persontype in the system"
          },
          "500" : {
            "description" : "Error creating persontype."
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
    "/api/v1/persontype/{id}" : {
      "get" : {
        "tags" : [ "PersonType" ],
        "summary" : "Gets a persontype entry.",
        "description" : "Gets a persontype entry.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The ID of the entry.",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "schema" : {
              "$ref" : "#/definitions/PersonType"
            }
          },
          "500" : {
            "description" : "error getting persontype with id 'id'"
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "PersonType" ],
        "summary" : "Updates a persontype entry\n",
        "description" : "Updates a persontype entry.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The ID of the entry.",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "name",
          "description" : "The type to update.",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/PersonType"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Updates the persontype."
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "PersonType" ],
        "summary" : "Deletes a persontype entry.\n",
        "description" : "Deletes a persontype entry.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The ID of the entry.",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Deletes the persontype.",
            "schema" : {
              "$ref" : "#/definitions/PersonType"
            }
          },
          "500" : {
            "description" : "Error deleting the persontype \"ID\"."
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "basicAuth" : {
      "description" : "HTTP Basic Authentication. Works over `HTTP` and `HTTPS`.",
      "type" : "basic"
    }
  },
  "definitions" : {
...
    "PersonType" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        }
      }
    },
...
    "FacultyRequest" : {
      "type" : "object",
      "properties" : {
        "notes" : {
          "type" : "string"
        }
      }
    },
    "AbuseRequest" : {
      "type" : "object",
      "properties" : {
        "item_type" : {
          "type" : "string"
        },
        "item_id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "item_url" : {
          "type" : "string"
        },
        "notes" : {
          "type" : "string"
        }
      }
    },
    "Verdict" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "approved" : {
          "type" : "boolean"
        },
        "verdict_notes" : {
          "type" : "string"
        }
      }
    },
...
  },
...
}